Expose injectable Hub server composition
